models-winning

+++ backend/api/internal/models/winning.go
@@ -4,7 +4,7 @@ import "time"
 
 type Winning struct {
     ID        uint      `gorm:"primaryKey"`
-    UserID    uint      `gorm:"not null"`
+    UserID    int64     `gorm:"not null"`  // Изменили uint на int64
     WinAmount float64   `gorm:"not null"`
     CreatedAt time.Time `gorm:"autoCreateTime"`
 }
